requirements.txt
alembic==0.9.4
captcha==0.3
celery==4.1.0
Flask==0.12.2
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
Jinja2==2.9.6
Pillow==4.2.1
PyMySQL==0.8.0
redis==2.10.5
SQLAlchemy==1.3.3

